refactor: separate Qt download workers from core
