+++ xl_updata_tool/docs/开发文档指南.md
@@ -34,7 +34,7 @@ xl_updata_tool/
 │   ├── __init__.py
 │   ├── core/               # 核心模块
 │   │   ├── __init__.py
-│   │   ├── downloader.py   # 下载器（含 DownloadWorker）
+│   │   ├── downloader.py   # HTTP、版本信息和校验函数（无 Qt）
 │   │   ├── database.py     # 数据库管理
 │   │   ├── version_manager.py  # 版本管理
 │   │   ├── bundle_parser.py    # 资源包解析（含 compute_delta）
@@ -61,6 +61,7 @@ xl_updata_tool/
 │       │   └── drag_list.py  # 支持拖放复制（DragListWidget）
 │       ├── workers/        # 后台工作线程（QThread）
 │       │   ├── __init__.py
+│       │   ├── download.py     # 下载/检查更新 Qt 线程包装
 │       │   ├── image_loader.py  # 缩略图异步加载（ImageLoadWorker）
 │       │   ├── preview_export.py  # 图片预览导出（PreviewExportWorker）
 │       │   ├── audio_decrypt.py  # 音频解密（AudioDecryptWorker）
@@ -96,15 +97,14 @@ xl_updata_tool/
 
 ## 核心模块说明
 
-### core/downloader.py
+### core/downloader.py 与 ui/workers/download.py
 
-提供下载相关功能：
+下载逻辑按依赖方向拆成两层：
 
-- `CheckUpdateThread` - 检查更新线程
-- `DownloadWorker` - 下载工作线程，支持：
-  - 信号：`progress`, `item_done`, `item_skip`, `item_fail`, `all_done`, `error`
-  - 自动重试机制（最多3次）
-  - MD5 校验
+- `app/core/downloader.py`：无 Qt 的 HTTP 请求、版本信息获取、Bundle 校验函数；
+- `app/ui/workers/download.py`：`CheckUpdateThread` 和 `DownloadWorker` 的 Qt 生命周期、进度信号和错误信号。
+
+`DownloadWorker` 保持原有信号兼容：`progress`、`item_done`、`item_skip`、`item_fail`、`all_done`、`error`；网络失败和 MD5 不匹配最多重试 3 次。
 
 ### core/database.py
 
@@ -113,16 +113,16 @@ SQLite 数据库管理，存储版本信息和资源包路径。
 ### core/logger.py
 
 日志系统，配置：
-- 固定日志文件：`logs/app.log`（使用 `RotatingFileHandler`）
+- 每次启动独立日志文件：`logs/app_YYYYMMDD_HHMMSS_ffffff.log`
 - 单文件最大 10 MB（`maxBytes=10*1024*1024`），超出时自动清空重写
-- 备份数量 0（`backupCount=0`），仅保留当前日志文件
+- 文件日志不可用时保留控制台日志，并输出降级警告
 - 文件日志级别 DEBUG，控制台日志级别 INFO
 - 编码 `utf-8`，支持中文字符
 - 同时输出到控制台和文件
 
 ### core/character_loader.py
 
-角色 Lua 数据解析引擎（无 Qt 依赖，纯模块级函数）。从 `data/material/assets/lua/` 解密后的 Lua 文件中提取结构化角色数据，供主窗口 `_load_character_data` 调用。主要函数：
+角色 Lua 数据解析引擎（无 Qt 依赖，纯模块级函数）。从 `output/lua/` 解密后的 Lua 文件中提取结构化角色数据，供主窗口 `_load_character_data` 调用。主要函数：
 - `parse_word_file(file_path)` - 解析 `BaseWord_cn.lua` 文本映射（id → 文本）
 - `parse_cv_file(file_path)` - 解析配音 `BaseCvNameCn.lua`
 - `parse_level_up_file(file_path)` - 解析 `BaseCardLevelUp.lua` 升级属性
@@ -160,14 +160,14 @@ SQLite 数据库管理，存储版本信息和资源包路径。
 
 ### ui/main_window.py
 
-主窗口控制器（重构后瘦身至 ~800 行），仅保留：
+主窗口控制器目前仍是较大的协调层，正在按领域逐步收口；目标是只保留：
 - 窗口骨架创建与初始化
 - 工具栏按钮事件分发
 - 版本列表展示与下载管理
 - 视图切换逻辑（版本列表 / 图片预览 / 音频管理器 / 角色数据）
 - 子模块导入与协调
 
-所有后台工作线程、对话框、自定义控件、FGUI 解析工具、Spine/FFmpeg 适配器均已拆分为独立模块（见下方子目录结构说明）。
+后台工作线程、对话框、自定义控件、FGUI 解析工具、Spine/FFmpeg 适配器已拆分为独立模块；版本/下载之外的主窗口领域协调逻辑仍按后续阶段继续迁移（见 `架构与协作基线.md`）。
 
 ### ui/ 子目录职责说明
