+++ xl_updata_tool/docs/架构与协作基线.md
@@ -1,0 +1,137 @@
+# XL Update Tool 架构与协作基线
+
+> 本文是稳定性与架构收口的基线文档。它描述当前可依赖的边界、运行时目录契约和后续拆分方向；代码与本文不一致时，先在 Issue 中说明，再同步修正文档。
+
+## 1. 项目定位与演进原则
+
+XL Update Tool 的主链路是：
+
+```text
+检查版本 → 计算 Bundle 差异 → 下载并校验 → 导入 AssetStudio → 分类导出 → 预览/角色/音频消费产物
+```
+
+项目已经从单体窗口演进到 `core / ui / workers / features / adapters` 的结构。下一阶段不追求一次性“大重构”，而是遵循四条原则：
+
+1. 先稳定契约，再拆文件；
+2. 纯逻辑可以在无 Qt、无 GUI、无真实游戏资源的环境中测试；
+3. 每个阶段都保持旧流程可回退、可验证；
+4. 共享文件改动小步提交，避免多人长期占用同一文件。
+
+## 2. 当前边界（As-Is）
+
+### `app/core/`
+
+负责可复用的纯逻辑：路径、日志、数据库、Bundle 解析/修复、角色 Lua 解析、网络请求和下载协议函数。
+
+约束：
+
+- 不得导入 `PySide6`、窗口、控件或信号；
+- 不得直接操作 UI 状态；
+- 外部工具调用应逐步收敛到 adapter/service 层；
+- 新增公共函数优先提供类型注解和最小单元测试。
+
+`app/core/downloader.py` 现在只保留 HTTP、版本信息和校验函数；Qt 线程包装位于 `app/ui/workers/download.py`。
+
+### `app/ui/`
+
+负责 Qt 生命周期、信号槽、用户提示和视图状态。UI 可以调用 core，但不应把文件遍历、解析规则和导出策略继续堆进 `main_window.py`。
+
+当前 `main_window.py` 仍是主要收口点，包含版本列表、下载、导入入口、视图切换和多个领域协调逻辑。它是后续拆分的重点，但拆分必须按领域逐批进行，不能与功能改动混在一起。
+
+### `workers/`
+
+负责耗时任务的 Qt 包装：线程生命周期、取消、进度信号和结果信号。耗时任务内部的纯计算应下沉到 core/features，Worker 不应成为新的业务单体。
+
+### `features/` 与 `adapters/`
+
+- `features/`：领域流程和可测试的业务编排，例如导出分类、角色数据、FairyGUI 解析；
+- `adapters/`：AssetStudio、SpineViewer、FFmpeg 等外部工具的命令构造、输出解析和兼容处理。
+
+## 3. 运行时目录与输出契约
+
+所有运行时路径通过 `app/core/path_utils.py` 获取，避免在业务代码中重复拼接项目根路径。
+
+| 区域 | 当前职责 | 长期规则 |
+|---|---|---|
+| `data/bundles/` | 原始下载 Bundle，按版本/哈希组织 | 下载产物只追加或替换临时文件后原子落盘 |
+| `data/material/` | AssetStudio 中间导出，兼容旧流程 | 视为可重建缓存，不作为长期展示数据唯一来源 |
+| `data/seeds/` | 版本分类 JSON 缓存 | 可重新拉取，缺失时自动恢复 |
+| `data/xl_updata.db` | 版本和 Bundle 元数据 | 磁盘扫描结果用于校准，不盲信历史状态 |
+| `output/lua/` | 反编译后的 Lua 展示输入 | 角色解析只读这里 |
+| `output/character_data/` | 角色结构化缓存 | 源文件变化时失效重建 |
+| `output/character/`、`output/fgui/`、`output/audio/`、`output/video/` | 可消费的分类产物 | 失败任务不得用半成品覆盖已成功产物 |
+| `logs/` | 启动级日志文件 | 文件日志不可用时仍保证控制台可诊断 |
+
+当前仍有部分预览/音频流程读取 `data/material/` 作为输入，这是兼容现实而不是理想边界。后续每迁移一个消费者，都必须同时保留旧路径回退或明确给出迁移提示，避免影响已有建设。
+
+## 4. 输出状态模型（目标契约）
+
+每个导出任务建议按以下状态推进：
+
+```text
+发现输入 → 写入临时目录 → 校验数量/格式 → 原子移动到 output → 写入完成标记
+       ↘ 失败：保留旧产物，清理临时目录，记录可定位错误
+```
+
+最低要求：
+
+- 单个文件失败不应让整批任务无故中止；
+- 任务取消后不能把未完成文件报告为成功；
+- 进度必须区分“已跳过、已成功、失败、取消”；
+- 增量任务按输入 hash/mtime 判断，不以“目录非空”作为全局完成条件；
+- 文件名或相对路径冲突必须显式报告，不能静默覆盖。
+
+## 5. 多人协作拆分策略
+
+领域认领建议如下：
+
+| 领域 | 主要目录 | 协作规则 |
+|---|---|---|
+| 版本/下载 | `core/downloader.py`、`ui/workers/download.py` | 先保持信号兼容，再改数据模型 |
+| 导入 AssetStudio | `ui/workers/import_as.py`、`ui/adapters/` | 任何输出目录改动必须同步契约文档 |
+| 图片/Spine | `ui/features/`、`ui/adapters/spine_adapter.py` | 先抽纯路径/配对逻辑，再动 UI |
+| 音频 | `ui/workers/audio_decrypt.py`、`core/album_map.py` | 增量解密和失败恢复单独验收 |
+| Lua/角色 | `core/character_loader.py`、`ui/views/character_view.py` | 解析器保持无 Qt，样例数据脱敏后再进测试 |
+| 主窗口 | `ui/main_window.py` | 仅做协调；大改必须单独 Issue/PR |
+
+推荐的 PR 粒度是“一个边界、一个可验证结果”，例如“下载 Qt 包装迁移”或“定义导出状态”，不要把 UI 美化、路径迁移和解析器重写放在同一 PR。
+
+## 6. 分阶段路线
+
+### P0：基线（本阶段）
+
+- 核心层去 Qt 依赖；
+- 建立 pytest、ruff、compile smoke 和 Windows CI；
+- 日志在受限目录下仍可诊断；
+- 明确安全规则、凭据规则和输出契约。
+
+### P1：稳定性
+
+- 下载写临时文件、校验通过后原子移动；
+- 导入/导出按任务临时目录执行，失败保留旧产物；
+- 将“目录非空即完成”改成逐文件状态；
+- 为增量导出、取消、重试和冲突补测试。
+
+### P2：架构拆分
+
+- 从 `main_window.py` 先抽版本/下载协调器；
+- 再抽导入 AssetStudio、音频、预览、角色四个领域协调器；
+- 让视图只负责展示，Worker 只负责线程边界；
+- 清理或隔离仍可触达的旧 `asset_browser.py` 流程。
+
+### P3：发布与长期维护
+
+- 统一配置注入外部工具路径、CDN 和运行目录；
+- 评估 Git LFS/发布附件，避免仓库历史继续膨胀；
+- 增加脱敏样例、回归资源清单和版本迁移说明；
+- 再决定便携版、安装包和工具运行时的打包策略。
+
+## 7. 每次改动的验收清单
+
+- 是否先有对应 Issue，且写明背景、边界和验收标准？
+- 是否影响公共文件或输出目录契约？
+- 是否有至少一个自动化测试或可复现的 smoke check？
+- 是否运行 `pytest`、`ruff --no-cache`、全量 Python compile？
+- 是否同步 README/开发文档/协作说明？
+- 是否能在失败、取消、重复执行后恢复？
+- 是否没有把密钥、真实账号、外部服务凭据或敏感资源写入仓库、Issue、PR、聊天和命令行参数？
