+++ 开发计划与沟通事项.md
@@ -4,6 +4,14 @@
 > 协作规范见 [CONTRIBUTING.md](CONTRIBUTING.md)。
 > 所有行号对应 `app/ui/main_window.py`（6356 行），随重构可能变化，以方法名为准。
 
+## 当前架构收口状态（2026-08）
+
+- 已建立 Issue #21，作为稳定性、架构边界和多人协作基线。
+- `app/core/downloader.py` 已去除 Qt 线程依赖，Qt 包装迁移到 `app/ui/workers/download.py`。
+- 已加入 pytest、ruff、全量 compile smoke、Windows CI 和核心层无 Qt 回归测试。
+- 已补充 [架构与协作基线](xl_updata_tool/docs/架构与协作基线.md)，后续以“先稳定契约、再按领域拆分”为原则。
+- `main_window.py`、`character_loader.py`、`asset_browser.py` 等大文件仍需渐进拆分；本轮不以一次性重写为目标。
+
 ---
 
 ## 一、已完成但需告知的事项
